@@ -23,22 +23,21 @@
         }
     ],
     "require": {
-        "php": "^7.4 || ^8.0",
-        "nezamy/route": "^1.2",
-        "symfony/translation": "^5.4",
+        "php": "^8.1",
+        "nezamy/route": "^1.3",
+        "symfony/translation": "^6.0",
         "nette/forms": "^3.1",
-        "parsedown/parsedown": "^1.7",
-        "emojione/emojione": "^4.5",
+        "erusev/parsedown": "^1.7",
+        "spatie/emoji": "^3.0",
         "tracy/tracy": "^2.9",
         "nette/robot-loader": "^3.4",
         "plasticbrain/php-flash-messages": "^1.0",
-        "leblanc-simon/parsedown-checkbox": "^0.0.4",
-        "nette/mail": "^3.1",
-        "latte/latte": "^2.11",
+        "nette/mail": "^4.0",
+        "latte/latte": "^3.0",
         "geertw/ip-anonymizer": "^1.1",
         "cbschuld/browser.php": "^1.9",
-        "ezimuel/php-secure-session": "^2.0",
-        "php-curl-class/php-curl-class": "^9.0"
+        "symfony/http-foundation": "^6.0",
+        "php-curl-class/php-curl-class": "^9.18"
     },
     "autoload": {
         "classmap": [
@@ -48,7 +47,7 @@
     "config": {
         "optimize-autoloader": true,
         "platform": {
-            "php": "7.4"
+            "php": "8.1"
         }
     },
     "prefer-stable": true
