@@ -16,7 +16,8 @@
         }
     ],
     "require": {
-        "php": ">=7.1"
+        "php": ">=7.2",
+        "ext-iconv": "*"
     },
     "provide": {
         "ext-mbstring": "*"
