@@ -15,16 +15,17 @@
 		}
 	],
 	"require": {
-		"php": ">=7.2 <8.3",
+		"php": "8.0 - 8.4",
 		"ext-session": "*",
 		"ext-json": "*"
 	},
 	"require-dev": {
-		"nette/utils": "^3.0",
+		"nette/utils": "^3.0 || ^4.0",
 		"nette/di": "^3.0",
-		"nette/mail": "^3.0",
+		"nette/http": "^3.0",
+		"nette/mail": "^3.0 || ^4.0",
 		"nette/tester": "^2.2",
-		"latte/latte": "^2.5",
+		"latte/latte": "^2.5 || ^3.0",
 		"psr/log": "^1.0 || ^2.0 || ^3.0",
 		"phpstan/phpstan": "^1.0"
 	},
@@ -42,7 +43,7 @@
 	},
 	"extra": {
 		"branch-alias": {
-			"dev-master": "2.9-dev"
+			"dev-master": "2.10-dev"
 		}
 	}
 }
