@@ -1,5 +1,15 @@
 <?php
 
-    $t->addResource('array', [
-        
-    ], 'en_EN'); 
+// src/translations/en_EN.php
+return [
+    'Login' => 'Login',
+    'Logout' => 'Logout',
+    'Password' => 'Password',
+    'Username' => 'Username',
+    'Settings' => 'Settings',
+    'Search' => 'Search',
+    'Edit' => 'Edit',
+    'Remove' => 'Remove',
+    'Add' => 'Add',
+    // You should copy all other 'Key' => 'Value', lines from your old file here
+];
